x-api changes, version detection changes
